backend/ms-llm-orchestrator/pyproject.toml
[project]
name = "ms-llm-orchestrator"
version = "0.1.0"
description = "A FastAPI service for LLM, RAG, and other generative AI tasks."
dependencies = [
    "fastapi",
    "uvicorn",
    "asyncpg",
    "pydantic-settings",
    "langchain-openai",
    "langchain-mcp-adapters",
    "langchain",
    "langchain_core",
    "langfuse",
    "utils",
]
requires-python = ">=3.11"

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.uv.sources]
utils = { path = "../utils", editable = true }

[tool.uv]
